mathy compatible with glm

+++ src/engine/source/yume_kernel.hpp
@@ -25,7 +25,7 @@ namespace yumegl {
         glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
         glfwWindowHint(GLFW_OPENGL_DEBUG_CONTEXT, true);
 
-        _window = glfwCreateWindow(WINDOW_WIDTH, WINDOW_HEIGHT, title.c_str(), nullptr, nullptr);
+        _window = glfwCreateWindow((int)WINDOW_WIDTH, (int)WINDOW_HEIGHT, title.c_str(), nullptr, nullptr);
 
         if (_window == nullptr) {
             std::cerr << "Failed to create GLFW window" << std::endl;
